refactor(embedded): #61 Phase 1 — carve mfsk-app-shared out of m5stack-s3-app
Move board-agnostic app logic (QSO FSM, time sync, TX picker, SNR
norm, NVS boot mode, log fanout, WiFi + UDP, UI data structures +
embedded-graphics draw routines, ADIF / flash log placeholders)
into a new `embedded-poc/mfsk-app-shared/` crate so a future
`m5stack-core2-app` (Phase 2) can reuse it. Pure rename + import
rewires — no behavior change. boot_mode::determine() now takes a
board-supplied KEY1 pin so the shared crate stays HW-agnostic.

decode_pipeline.rs stays in the s3-app bin since it owns the
heap_caps BASIS alloc; carve only if Phase 2 shows a clean cut.

Verified on hardware (S3 LX7, BootMode::Wifi):
- cargo build --release: clean, 32.85 s
- flash + 90 s monitor: 500+ alive ticks, heap flat at
  8,534,168 free / 184,643 internal / 110,592 largest; no
  panic/abort/E (esp_idf) lines.
