+++ embedded-poc/mfsk-app-shared/src/adif.rs
