+++ embedded-poc/mfsk-app-shared/src/lib.rs
@@ -1,0 +1,19 @@
+//! Board-agnostic logic for the mfsk FT8 controller apps.
+//!
+//! See `Cargo.toml` for the shared/board boundary rationale. Each module
+//! below is moved verbatim from the original `m5stack-s3-app` crate
+//! (see `git log --follow` on each file for pre-extraction history).
+
+#![allow(dead_code)]
+
+pub mod adif;
+pub mod boot_mode;
+pub mod flash_log;
+pub mod log_sink;
+pub mod qso;
+pub mod snr_norm;
+pub mod time_sync;
+pub mod tx_picker;
+pub mod udp_log;
+pub mod ui;
+pub mod wifi;
