+++ embedded-poc/mfsk-app-shared/src/log_sink.rs
