+++ embedded-poc/mfsk-app-shared/src/qso.rs
