+++ embedded-poc/mfsk-app-shared/src/snr_norm.rs
