+++ embedded-poc/mfsk-app-shared/src/ui/menu.rs
