+++ embedded-poc/mfsk-app-shared/src/ui/state.rs
