+++ embedded-poc/mfsk-app-shared/src/ui/status_bar.rs
