+++ embedded-poc/mfsk-app-shared/src/wifi.rs
